+++ react-ts-template/src/App.tsx
@@ -1,9 +1,10 @@
 import type { RouteObject } from "react-router-dom";
 import { Outlet, Link, useRoutes, useParams } from "react-router-dom";
-import Header from "./components/Header";
+import Header from "./components/Header/Header";
+import Footer from "./components/Footer/Footer";
 
 export default function App() {
     return (
-        <Header />
+        <><Header /><Footer /></>
     );
 }
