+++ react-ts-template/src/assets/styles/theme.css
@@ -1,4 +1,5 @@
 :root {
     --warning: #FFC832;
     --danger: rgba(237,26,59,1);
+    --dark-grey: #555;
   }
