+++ react-ts-template/src/components/Footer/Footer.module.css
@@ -1,0 +1,32 @@
+@import "../../assets/styles/theme.css";
+
+.main_container {
+    position: absolute;
+    width: 100%;
+    height: auto;
+    background-color: var(--dark-grey);
+    display: flex;
+    align-items: left;
+    flex-direction: column;
+    flex-wrap: wrap;
+    bottom: 0;
+}
+
+.logo {
+    position: relative;
+    width: 100px;
+    height: 40px;
+    background: url(../../assets/images/marvel_logo.svg)
+    no-repeat center;
+    background-size: cover;
+    margin-left: 20px;
+    margin-top: 20px;
+}
+
+.text {
+    position: relative;
+    color: white;
+    font-size:16px;
+    margin-left: 20px;
+    margin-top: 5px;
+}
